RocksDB Revert Function back to be used purely for ConditionExpression and not UpdateExpression
